managing preflight 1

+++ controllers/ticketsController.js
@@ -2,6 +2,8 @@ const { Ticket } = require('../models')
 
 class TicketsController {
   async create(req, res, next) {
+    if (req.method === 'OPTIONS') return next(res.status(200))
+
     try {
       let { creator, description, issue, problem, projectid, severity, solution, status } = req.body
       const ticketData = { description, issue, problem, projectid, severity, solution, status }
@@ -20,6 +22,8 @@ class TicketsController {
   }
 
   async update(req, res, next) {
+    if (req.method === 'OPTIONS') return next(res.status(200))
+
     try {
       const ticket = req.body
       const { id } = ticket
@@ -32,6 +36,8 @@ class TicketsController {
   }
 
   async delete(req, res, next) {
+    if (req.method === 'OPTIONS') return next(res.status(200))
+
     try {
       const id = Number(req.body.id)
       await Ticket.destroy({ where: { id } })
@@ -43,6 +49,8 @@ class TicketsController {
   }
 
   async getAll(req, res, next) {
+    if (req.method === 'OPTIONS') return next(res.status(200))
+
     try {
       const { id } = req.params
       const tickets = await Ticket.findAll({
